use 2018-style modules
